Changed code organization
